I3DC: Interactive Three-Dimensional Cubes (ICDE 2009)
I3DC: Interactive Three-Dimensional Cubes

We present the I3DC system prototype, which constructs the cube for tens
of millions of data items within milliseconds, and provides high-quality
cube visualization as well as highly-interactive OLAP operations. Our
approach is based on a novel blending-as-aggregation (BAA) algorithm
that maps distributive OLAP aggregations to the intrinsic rendering
mechanisms of the GPU. Our system runs entirely on the GPU and requires
no precomputations.